feat: moved qrcanbox to molecules
