+++ src/features/components/molecules/qr-scan-box/styles.module.scss
@@ -2,7 +2,3 @@
     border: 0.5px solid var(--color-white);
     border-radius: 24px;
 }
-
-.download_col {
-    max-inline-size: 416px;
-}
