+++ src/features/pages/p2p/banner/styles.module.scss
@@ -24,3 +24,7 @@
         gap: 24px;
     }
 }
+
+.download_col {
+    max-inline-size: 416px;
+}
